+++ bakery/apps.py
@@ -1,8 +1,9 @@
 import logging
 
-import fs
 from django.apps import AppConfig
 from django.conf import settings
+
+from bakery.filesystem import RootedFilesystem
 
 logger = logging.getLogger(__name__)
 
@@ -11,4 +12,4 @@ class BakeryConfig(AppConfig):
     name = "bakery"
     verbose_name = "Bakery"
     filesystem_name = getattr(settings, "BAKERY_FILESYSTEM", "osfs:///")
-    filesystem = fs.open_fs(filesystem_name)
+    filesystem = RootedFilesystem.from_url(filesystem_name)